# swagger.yaml
openapi: 3.0.0
info:
  title: Applicant Dashboard API
  description: |
    # Applicant Dashboard API Documentation
    
    This API provides access to the Applicant Dashboard system, allowing applicants to track their application status, 
    communicate with the recruitment team, and manage their application details.
    
    ## Key Features
    - Magic link authentication for secure access
    - Application status tracking
    - Two-way communication
    - Personalized insights and updates
  version: 1.0.0
  contact:
    name: ProDG Support Team
    email: support@prodg.studio
    url: https://prodg.studio/support
  license:
    name: Proprietary
    url: https://prodg.studio/terms
servers:
  - url: /api
    description: Production API server
  - url: http://localhost:3000/api
    description: Local development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: applicant_session
      description: HTTP-only cookie containing the session token for authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message describing what went wrong
      required:
        - error
      example:
        error: "Application not found"
...
    MessageType:
      type: string
      enum:
        - message
        - status_update
        - request
        - notification
      description: Type of communication message
    SenderType:
      type: string
      enum:
        - applicant
        - admin
        - system
      description: Who sent the message
    Communication:
      type: object
      properties:
        _id:
          type: string
          description: Unique message identifier
        applicationId:
          type: string
          description: Associated application ID
        senderId:
          type: string
          description: ID of the sender
        senderType:
          $ref: '#/components/schemas/SenderType'
        message:
          type: string
          description: Message content
          maxLength: 5000
        attachments:
          type: array
          items:
            type: string
            format: uri
          description: List of attachment URLs
        timestamp:
          type: string
          format: date-time
          description: When the message was sent
        isRead:
          type: boolean
          description: Whether the message has been read by the recipient
        messageType:
          $ref: '#/components/schemas/MessageType'
      required:
        - _id
        - applicationId
        - senderId
        - senderType
        - message
        - timestamp
        - isRead
        - messageType
      example:
        _id: "60b2c5f9d8a2e45fa7c12345"
        applicationId: "60a3d1e9b50c2b3a24c8e4f2"
        senderId: "admin123"
        senderType: "admin"
        message: "Thank you for your application! We're currently reviewing it and will get back to you soon."
        timestamp: "2023-06-01T09:00:00Z"
        isRead: false
        messageType: "message"
    CommunicationList:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Communication'
          description: List of communication messages
        hasMore:
          type: boolean
          description: Whether there are more messages available
        unreadCount:
          type: integer
          description: Number of unread messages
      required:
        - messages
        - hasMore
      example:
        messages:
          - _id: "60b2c5f9d8a2e45fa7c12345"
            applicationId: "60a3d1e9b50c2b3a24c8e4f2"
            senderId: "admin123"
            senderType: "admin"
            message: "Thank you for your application!"
            timestamp: "2023-06-01T09:00:00Z"
            isRead: true
            messageType: "message"
          - _id: "60b2c5f9d8a2e45fa7c12346"
            applicationId: "60a3d1e9b50c2b3a24c8e4f2"
            senderId: "system"
            senderType: "system"
            message: "Your application status has been updated to 'reviewing'."
            timestamp: "2023-06-02T10:00:00Z"
            isRead: false
            messageType: "status_update"
        hasMore: false
        unreadCount: 1
...
    NewMessage:
      type: object
      properties:
        message:
          type: string
          description: Message content
          maxLength: 5000
      required:
        - message
      example:
        message: "Hello, I have a question about my application. When can I expect to hear back about the next steps?"
paths:
...
  /applicant/communications:
    get:
      summary: Get messages
      description: |
        Retrieves the applicant's messages.
        
        This endpoint returns a list of communication messages between the applicant and the recruitment team.
        Messages are sorted by timestamp in descending order (newest first).
        
        Pagination is supported through the 'limit' and 'before' parameters.
      tags:
        - Communication
      operationId: getMessages
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          description: Maximum number of messages to retrieve
        - in: query
          name: before
          schema:
            type: string
            format: date-time
          description: Get messages before this timestamp
          example: "2023-06-01T12:00:00Z"
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunicationList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          examples:
            unauthorized:
              value:
                error: "Unauthorized"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Send message
      description: |
        Sends a message from the applicant to the recruitment team.
        
        This endpoint allows applicants to send a text message to communicate with the recruitment team.
        Messages are limited to 5000 characters.
        
        When a message is sent successfully, it's added to the communication history and the recruitment team
        is notified based on their notification settings.
      tags:
        - Communication
      operationId: sendMessage
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMessage'
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Communication'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          examples:
            emptyMessage:
              value:
                error: "Message cannot be empty"
            messageTooLong:
              value:
                error: "Message exceeds maximum length of 5000 characters"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          examples:
            unauthorized:
              value:
                error: "Unauthorized"
        '403':
          description: Communication disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          examples:
            communicationDisabled:
              value:
                error: "Communication is currently disabled for this application"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 